tentativa-fa-db/pyproject.toml
[tool.poetry]
name = "banco-materia-full"
version = "0.1.0"
description = ""
authors = ["MatheusFariasSousa <143126173+MatheusFariasSousa@users.noreply.github.com>"]
license = "a"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pytz = "^2024.1"
tzdata = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
